Fix potential use of uninitialized variable.
It appears in some cases control might jump to the end of `fuzz_one` via
the `abandon_entry` label, where it gets freed, prior to `position_map`
initialization. Setting it to `NULL` turns this into a harmless
`free(NULL)`. Might manifest as a double free in some cases, probably
due to reuse of old stack space in a subsequent call.